+++ R/zzz.R
@@ -4,6 +4,7 @@ utils::globalVariables(
         "radius", "angle", "CI_upper", "CI_lower",
         "CI_lower_X", "CI_lower_Y", "CI_upper_X", "CI_upper_Y",
         "index", "X_new", "Y_new", "CI_X", "CI_Y", "se.fit",
-        ".index"
+        "frame_id", "chn", "coord", "coord_knot", "value",
+        "displ_id", "knot", "Date Time of recording"
     )
 )