+++ NEWS.md
@@ -1,6 +1,36 @@
+# rticulate v1.7.3.9000
+
+## BREAKING
+
+* `read_aaa()` now can read AAA data with existing headers, both fan-line and DeepLabCut data.
+
+  * Now any column starting with `X` or `Y` will be selected for pivoting when `format = "long"` (the default).
+
+* This enhancement has brought in a few breaking changes:
+
+  * The order of the arguments has changed.
+  * The argument `fan_line` doesn't have a default any more, so it has to be specified.
+
+## Added
+
+* `read_ag500_pos()` to read `.pos` files from Carstens AG500 electro-magnetic-articulographer.
+
+* `filter_signal()` to filter a signal with a Savitzky-Golay or Butterworth filter.
+
+* `resample_signal()` to up/down-sample a signal using interpolation.
+
+* `get_velocity()` and `get_acceleration()` to calculate the velocity and acceleration (first and second derivative) of displacement.
+
+* `get_landmarks()` to get several gestural landmarks from the displacement of a single gesture.
+
+
+
+
+
 # rticulate v1.7.4
 
 * Fix CRAN note about language field.
+
 
 
 # rticulate v1.7.3